requirements.txt
numpy==1.26.4
pandas==2.2.2
opencv-python==4.10.0.84
scikit-image==0.24.0
scikit-learn==1.5.1
tqdm==4.66.5
Pillow==10.4.0
PyYAML==6.0.2
torch==2.3.1
torchvision==0.18.1
kagglehub==0.3.5
joblib==1.4.2
